+++ packages/markdown/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./core"
 export * from "./ast"
 export * from "./markers"
-export * from "./wikilink"